+++ zida-module-temu/src/main/java/com/zida/cbec/module/temu/controller/admin/productspu/vo/TemuProductSpuRespVO.java
@@ -39,7 +39,7 @@ public class TemuProductSpuRespVO {
 
     @Schema(description = "市场价格", example = "30916")
     @ExcelProperty("市场价格")
-    private BigDecimal marketPrice;
+    private Long marketPrice;
 
     @Schema(description = "商品售价", example = "17781")
     @ExcelProperty("商品售价")
